moves animations and trackByFns files to shared folder root
