+++ apps/menu-matriarch/src/app/features/meals/meal-details/meal-details.component.ts
@@ -7,7 +7,7 @@ import { MealService } from '@services/meal.service';
 import { UserService } from '@services/user.service';
 import { trackBySelf } from '@atocha/core/ui';
 import { getDishTypes } from '@atocha/menu-matriarch/types';
-import { dishTrackByFn } from '@utility/domain/track-by-functions';
+import { dishTrackByFn } from '@shared/track-by-functions';
 
 @Component({
   selector: 'app-meal-details',
