+++ apps/menu-matriarch/src/app/features/menus/menu-card/menu-card.component.ts
@@ -5,7 +5,7 @@ import { BehaviorSubject } from 'rxjs';
 import { Day, MenuEntry, Orientation } from '@atocha/menu-matriarch/types';
 import { MenuService } from '@services/menu.service';
 import { PrintService } from '@services/print.service';
-import { menuEntryTrackByFn } from '@utility/domain/track-by-functions';
+import { menuEntryTrackByFn } from '@shared/track-by-functions';
 
 type State = 'default' | 'renaming' | 'changingStartDay';
